+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix graphicCorner complication gauge range text reversed
 - Fix graphicCircular complication with open gauge
 - Fix graphicRectangular complication place holder body1TextProvider
+- Add graphicExtraLarge complication for watchOS 7
 
 ## [watchOS 1.1.0(9)] - 2019-09-25
 - Remove start timer action